获取window的decorview设置展示样式,然后设置window的statusbar的颜色为透明  Color.TRANSPARENT

+++ app/src/main/java/com/kunkun/sunnyweather/logic/model/DailyResponse.kt
@@ -94,7 +94,7 @@ data class DailyResponse(
         val avg: Double,
         val date: String,
         val max: Double,
-        val min: Int,
+        val min: Double,
         val probability: Int
     )
